migrate twig to separate class
