+++ src/Common/Twig/TwigExtension.php
@@ -3,16 +3,16 @@
 /**
  * TwigExtension class.
  *
- * OpenEMR central interface for twig.
+ * OpenEMR central extension interface for twig.
  *
  * @package   OpenEMR
  * @link      https://www.open-emr.org
  * @author    Brady Miller <brady.g.miller@gmail.com>
- * @copyright Copyright (c) 2019 Brady Miller <brady.g.miller@gmail.com>
+ * @copyright Copyright (c) 2019-2021 Brady Miller <brady.g.miller@gmail.com>
  * @license   https://github.com/openemr/openemr/blob/master/LICENSE GNU General Public License 3
  */
 
-namespace OpenEMR\Core;
+namespace OpenEMR\Common\Twig;
 
 use OpenEMR\Core\Header;
